Tables and exceptions relation is added.

+++ src/main/java/com/maen/hotel/service/AddressService.java
@@ -1,11 +1,13 @@
 package com.maen.hotel.service;
 
+import com.maen.hotel.exceptions.GlobalNotFoundException;
 import com.maen.hotel.persistence.entity.AddressEntity;
 import com.maen.hotel.persistence.repository.IAddressRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @AllArgsConstructor
@@ -17,11 +19,23 @@ public class AddressService {
         return this.addressRepository.findAll();
     }
 
+    public Optional<AddressEntity> getByIdAddress(Integer id) throws GlobalNotFoundException {
+        Optional<AddressEntity> address = addressRepository.findById(id);
+        if (!address.isPresent()){
+            throw new GlobalNotFoundException("Resource not found.");
+        }
+        return this.addressRepository.findById(id);
+    }
+
     public AddressEntity saveAddrees(AddressEntity addressEntity){
         return this.addressRepository.save(addressEntity);
     }
 
-    public AddressEntity updateAddress(Integer id, AddressEntity addressEntity){
+    public AddressEntity updateAddress(Integer id, AddressEntity addressEntity) throws GlobalNotFoundException{
+        Optional<AddressEntity> update = addressRepository.findById(id);
+        if (!update.isPresent()){
+            throw new GlobalNotFoundException("Update failed, resource not found.");
+        }
         AddressEntity address = addressRepository.findById(id).get();
         address.setCountry(addressEntity.getCountry());
         address.setCity(addressEntity.getCity());
@@ -32,7 +46,11 @@ public class AddressService {
         return address;
     }
 
-    public void deleteAddress(Integer id){
+    public void deleteAddress(Integer id) throws GlobalNotFoundException{
+        Optional<AddressEntity> delete = addressRepository.findById(id);
+        if (!delete.isPresent()){
+            throw new GlobalNotFoundException("Error when trying to delete a non-existing resource.");
+        }
         this.addressRepository.deleteById(id);
     }
 }
